Melhora a injeção de dependência e validação de dados
- Adiciona construtores em App e TransactionList para aceitar instâncias específicas, melhorando a injeção de dependência.
- Registra páginas como serviços transitórios em MauiProgram.
- Altera tipo de Value de Decimal para Double em Transaction.
- Adiciona recursos de estilo em TransactionList.xaml.
- Implementa exibição de mensagens de erro em TransactionNew.
- Melhora a validação de entrada em TransactionNew.

+++ Financeiro/Models/Transaction.cs
@@ -9,5 +9,5 @@ public class Transaction
     public TransactionType Type { get; set; }
     public required String Name { get; set; }
     public DateTimeOffset Date { get; set; }
-    public Decimal Value { get; set; }
+    public Double Value { get; set; }
 }
